fix: prevent large worker backlogs
