+++ internal/services/dispatcher/server.go
@@ -33,182 +33,6 @@ import (
 	"github.com/hatchet-dev/hatchet/pkg/telemetry"
 )
 
-type subscribedWorker struct {
-	// stream is the server side of the RPC stream
-	stream contracts.Dispatcher_ListenServer
-
-	// finished is used to signal closure of a client subscribing goroutine
-	finished chan<- bool
-
-	sendMu sync.Mutex
-
-	workerId string
-}
-
-func (worker *subscribedWorker) StartStepRun(
-	ctx context.Context,
-	tenantId string,
-	stepRun *dbsqlc.GetStepRunForEngineRow,
-	stepRunData *dbsqlc.GetStepRunDataForEngineRow,
-) error {
-	ctx, span := telemetry.NewSpan(ctx, "start-step-run") // nolint:ineffassign
-	defer span.End()
-
-	inputBytes := []byte{}
-
-	if stepRunData.Input != nil {
-		inputBytes = stepRunData.Input
-	}
-
-	stepName := stepRun.StepReadableId.String
-
-	action := &contracts.AssignedAction{
-		TenantId:      tenantId,
-		JobId:         sqlchelpers.UUIDToStr(stepRun.JobId),
-		JobName:       stepRun.JobName,
-		JobRunId:      sqlchelpers.UUIDToStr(stepRun.JobRunId),
-		StepId:        sqlchelpers.UUIDToStr(stepRun.StepId),
-		StepRunId:     sqlchelpers.UUIDToStr(stepRun.SRID),
-		ActionType:    contracts.ActionType_START_STEP_RUN,
-		ActionId:      stepRun.ActionId,
-		ActionPayload: string(inputBytes),
-		StepName:      stepName,
-		WorkflowRunId: sqlchelpers.UUIDToStr(stepRun.WorkflowRunId),
-		RetryCount:    stepRun.SRRetryCount,
-		// NOTE: This is the default because this method is unused
-		Priority: 1,
-	}
-
-	if stepRunData.AdditionalMetadata != nil {
-		metadataStr := string(stepRunData.AdditionalMetadata)
-		action.AdditionalMetadata = &metadataStr
-	}
-
-	if stepRunData.ChildIndex.Valid {
-		action.ChildWorkflowIndex = &stepRunData.ChildIndex.Int32
-	}
-
-	if stepRunData.ChildKey.Valid {
-		action.ChildWorkflowKey = &stepRunData.ChildKey.String
-	}
-
-	if stepRunData.ParentId.Valid {
-		parentId := sqlchelpers.UUIDToStr(stepRunData.ParentId)
-		action.ParentWorkflowRunId = &parentId
-	}
-
-	worker.sendMu.Lock()
-	defer worker.sendMu.Unlock()
-
-	return worker.stream.Send(action)
-}
-
-func (worker *subscribedWorker) StartStepRunFromBulk(
-	ctx context.Context,
-	tenantId string,
-	stepRun *dbsqlc.GetStepRunBulkDataForEngineRow,
-) error {
-	ctx, span := telemetry.NewSpan(ctx, "start-step-run-from-bulk") // nolint:ineffassign
-	defer span.End()
-
-	inputBytes := []byte{}
-
-	if stepRun.Input != nil {
-		inputBytes = stepRun.Input
-	}
-
-	stepName := stepRun.StepReadableId.String
-
-	action := &contracts.AssignedAction{
-		TenantId:      tenantId,
-		JobId:         sqlchelpers.UUIDToStr(stepRun.JobId),
-		JobName:       stepRun.JobName,
-		JobRunId:      sqlchelpers.UUIDToStr(stepRun.JobRunId),
-		StepId:        sqlchelpers.UUIDToStr(stepRun.StepId),
-		StepRunId:     sqlchelpers.UUIDToStr(stepRun.SRID),
-		ActionType:    contracts.ActionType_START_STEP_RUN,
-		ActionId:      stepRun.ActionId,
-		ActionPayload: string(inputBytes),
-		StepName:      stepName,
-		WorkflowRunId: sqlchelpers.UUIDToStr(stepRun.WorkflowRunId),
-		RetryCount:    stepRun.SRRetryCount,
-		Priority:      stepRun.Priority,
-	}
-
-	if stepRun.AdditionalMetadata != nil {
-		metadataStr := string(stepRun.AdditionalMetadata)
-		action.AdditionalMetadata = &metadataStr
-	}
-
-	if stepRun.ChildIndex.Valid {
-		action.ChildWorkflowIndex = &stepRun.ChildIndex.Int32
-	}
-
-	if stepRun.ChildKey.Valid {
-		action.ChildWorkflowKey = &stepRun.ChildKey.String
-	}
-
-	if stepRun.ParentId.Valid {
-		parentId := sqlchelpers.UUIDToStr(stepRun.ParentId)
-		action.ParentWorkflowRunId = &parentId
-	}
-
-	worker.sendMu.Lock()
-	defer worker.sendMu.Unlock()
-
-	return worker.stream.Send(action)
-}
-
-func (worker *subscribedWorker) StartGroupKeyAction(
-	ctx context.Context,
-	tenantId string,
-	getGroupKeyRun *dbsqlc.GetGroupKeyRunForEngineRow,
-) error {
-	ctx, span := telemetry.NewSpan(ctx, "start-group-key-action") // nolint:ineffassign
-	defer span.End()
-
-	inputData := getGroupKeyRun.GetGroupKeyRun.Input
-	workflowRunId := sqlchelpers.UUIDToStr(getGroupKeyRun.WorkflowRunId)
-	getGroupKeyRunId := sqlchelpers.UUIDToStr(getGroupKeyRun.GetGroupKeyRun.ID)
-
-	worker.sendMu.Lock()
-	defer worker.sendMu.Unlock()
-
-	return worker.stream.Send(&contracts.AssignedAction{
-		TenantId:         tenantId,
-		WorkflowRunId:    workflowRunId,
-		GetGroupKeyRunId: getGroupKeyRunId,
-		ActionType:       contracts.ActionType_START_GET_GROUP_KEY,
-		ActionId:         getGroupKeyRun.ActionId,
-		ActionPayload:    string(inputData),
-	})
-}
-
-func (worker *subscribedWorker) CancelStepRun(
-	ctx context.Context,
-	tenantId string,
-	stepRun *dbsqlc.GetStepRunForEngineRow,
-) error {
-	ctx, span := telemetry.NewSpan(ctx, "cancel-step-run") // nolint:ineffassign
-	defer span.End()
-
-	worker.sendMu.Lock()
-	defer worker.sendMu.Unlock()
-
-	return worker.stream.Send(&contracts.AssignedAction{
-		TenantId:      tenantId,
-		JobId:         sqlchelpers.UUIDToStr(stepRun.JobId),
-		JobName:       stepRun.JobName,
-		JobRunId:      sqlchelpers.UUIDToStr(stepRun.JobRunId),
-		StepId:        sqlchelpers.UUIDToStr(stepRun.StepId),
-		StepRunId:     sqlchelpers.UUIDToStr(stepRun.SRID),
-		ActionType:    contracts.ActionType_CANCEL_STEP_RUN,
-		StepName:      stepRun.StepReadableId.String,
-		WorkflowRunId: sqlchelpers.UUIDToStr(stepRun.WorkflowRunId),
-		RetryCount:    stepRun.SRRetryCount,
-	})
-}
-
 func (s *DispatcherImpl) Register(ctx context.Context, request *contracts.WorkerRegisterRequest) (*contracts.WorkerRegisterResponse, error) {
 	tenant := ctx.Value("tenant").(*dbsqlc.Tenant)
 	tenantId := sqlchelpers.UUIDToStr(tenant.ID)
